Break out FilterBar component and rename types
