+++ src/js/types.ts
@@ -14,3 +14,18 @@ export interface Word {
   word: string;
   probability: number;
 }
+
+export type SearchField = 'episode' | 'title' | 'description';
+export type SearchFunction = (ep: Episode, searchTerm: string) => boolean;
+export type SearchFunctions = {
+  [k in SearchField]: SearchFunction;
+};
+
+export type FiltersState = {
+  [k in SearchField]: boolean;
+};
+
+export interface FilterBarProps {
+  filters: FiltersState;
+  onToggle: (args: { name: string; checked: boolean }) => void;
+}
